+++ src/main/java/net/frozen1753/copperequipments/mixin/PlayerEntityMixin.java
@@ -1,7 +1,9 @@
 package net.frozen1753.copperequipments.mixin;
 
+import net.frozen1753.copperequipments.util.accessor.ForcedOxidationFlagHolder;
 import net.frozen1753.copperequipments.util.accessor.ScrapeFlagHolder;
 import net.frozen1753.copperequipments.util.accessor.UnwaxFlagHolder;
+import net.frozen1753.copperequipments.util.accessor.WaxFlagHolder;
 import net.minecraft.entity.player.PlayerEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -9,9 +11,11 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(PlayerEntity.class)
-public class PlayerEntityMixin implements ScrapeFlagHolder, UnwaxFlagHolder {
+public class PlayerEntityMixin implements ScrapeFlagHolder, UnwaxFlagHolder, WaxFlagHolder, ForcedOxidationFlagHolder {
     private boolean playedScrapeThisTick = false;
     private boolean playedUnwaxThisTick = false;
+    private boolean playedWaxThisTick = false;
+    private boolean playedForcedOxidationThisTick = false;
 
     @Override
     public boolean hasPlayedScrapeThisTick() {
@@ -33,10 +37,32 @@ public class PlayerEntityMixin implements ScrapeFlagHolder, UnwaxFlagHolder {
         this.playedUnwaxThisTick = value;
     }
 
+    @Override
+    public boolean hasPlayedForcedOxidationThisTick() {
+        return playedForcedOxidationThisTick;
+    }
+
+    @Override
+    public void setPlayedForcedOxidationThisTick(boolean value) {
+        this.playedForcedOxidationThisTick = value;
+    }
+
+    @Override
+    public boolean hasPlayedWaxThisTick() {
+        return playedWaxThisTick;
+    }
+
+    @Override
+    public void setPlayedWaxThisTick(boolean value) {
+        this.playedWaxThisTick = value;
+    }
+
     @Inject(method = "tick", at = @At("HEAD"))
     private void resetFlags(CallbackInfo ci) {
         this.playedScrapeThisTick = false;
         this.playedUnwaxThisTick = false;
+        this.playedForcedOxidationThisTick = false;
+        this.playedWaxThisTick = false;
     }
 }