+++ src/main/java/net/frozen1753/copperequipments/recipe/custom/ModRecipe.java
@@ -1,0 +1,8 @@
+package net.frozen1753.copperequipments.recipe.custom;
+
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.world.World;
+
+public interface ModRecipe {
+    void playSound(World world, PlayerEntity player);
+}