+++ src/main/java/net/frozen1753/copperequipments/util/accessor/ForcedOxidationFlagHolder.java
@@ -1,0 +1,6 @@
+package net.frozen1753.copperequipments.util.accessor;
+
+public interface ForcedOxidationFlagHolder {
+    boolean hasPlayedForcedOxidationThisTick();
+    void setPlayedForcedOxidationThisTick(boolean value);
+}