+++ src/main/java/net/frozen1753/copperequipments/util/accessor/WaxFlagHolder.java
@@ -1,0 +1,6 @@
+package net.frozen1753.copperequipments.util.accessor;
+
+public interface WaxFlagHolder {
+    boolean hasPlayedWaxThisTick();
+    void setPlayedWaxThisTick(boolean value);
+}